Update simpleCalculator.js
Fixed an issue where after performing a calculation and continuing to perform more calculations, the code does not continue after the previous calculation causing the code to reset instead. So the issue was resolved by checking if variable 'operated' is true and if so set it to false to let the program know to continue on from the previous calculation.