Update simpleCalculator.js
Modified the reset() function to reset the 'screen' object to display '0' if the button displays 'C' by adding an if-statement. Otherwise if 'CE' is displayed, then the entire calculation is set to 0.

Also modified the removeDigit() function that deletes character one by one to check if 'scrren' object displays value zero '0', then display "CE" instead of "C".