Modified the reset() and percentage() functions
When playing with the "C" and "CE" buttons, I encountered times when "C" was clicked but didn't change to "CE". To fix this issue I changed the if-condition checking if button displays "CE". This way when ever the button is clicked by the user, the code line 'C_CE = 'CE'" is always executed while variable "a" remains the same whenever "C" is clicked.

Also Made changes to the percentage "%" button by leaving the button blue when clicked to indicate that the calculator will calculate percentage when clicking operate!