Update simpleCalculator.js
Modified the function inverse() to handle when an operation character (like '+','-','x' etc) is present in the screen object to avoid the 'screen' object displaying "NaN". I made 2 temporary variables "inversed" and "op" to contain the string value from 'screen' object and the operation character respectively. An if-statement is also made to check if the first character from string is operation character, if so then it extracts the operation character into 'op' and the string 'inversed' removes the operation character using slice().