Update simpleCalculator.js
Saw a bug when a friend of mine was playing around my digital calculator. The issue was that after square rooting a number and then selecting a digit, the digit gets added into the screen object along side the square rooted value when instead the 'screen' object should clear itself and display only the digit selected after the square root value.

The issue was fixed by adding a line of code 'operated = true;' in the square_root() function to tell the program a function has occurred, thus when the user selects a digit button after squre rooting a number, the program knows to reset the 'screen' object.

Also fixed an issue where after selecting the percentage button, it remains coloured blue after clicking on the equal "=" button. This issue was fixed by calling the function 'percentage()' in the 'reset()' and 'operating()' function using an if-statement to check if the variable is already 'false'. From now on, after calling the 'operating()' function, the percentage button gets reset back to orange colour.