Fixed display issues with the percentage button
When testing my percentage button on a laptop device, after clicking the percentage '%' button twice, the CSS property failed to work when hovering over. To fix this I made another class in CSS called '.btn-operators-clicked-on' which has the same properties as '.btn-operators' but its value for 'color' and 'background color' are different.

This way, instead of Javascript changes the button objects colour directly, it changes the class name back and forth.

I also fixed an issue where after inversing a number and selecting a digit, the 'screen' object just adds the selected digit character rather than clearing everything first before displaying the selected digit character. The issue was fixed by adding a code line 'operator = true;' just like fixing the issue with the percentage '%' button in a previous commit